requirements.txt
streamlit
numpy
matplotlib
qiskit
qiskit-aer
qiskit-algorithmsstreamlit
sympy

